ElizaOS plugin, for DAO purposes only.
The burntrust registry is a for onchain DAO purposes only specially for big projects that have rugged or scammed for millions. This isn't your standard pump fun scam registry. We understand the DAO is autonomous and not responsible if you're on this list or if others fork it and add to the list.